Change command function type to uint64 when receive from input
The client side is sending the value of array index in this field. The size of uint32 should be safe for an array index, but considering the definition in the structure, using uint64 is better. Note: To align the structure size, the client side also need to use uint64 when sending this array index.